added module to crawl exports
