+++ src/gary.js
@@ -2,8 +2,7 @@ const http = require("axios");
 const auth = require("../auth.json");
 const lib  = require("./lib.js");
 const queries  = require("./queries.js");
-
-const rego   = require("./regocrawl.js");
+const crawls   = require("./crawls.js");
 
 exports.feats = [
 
@@ -94,10 +93,9 @@ exports.feats = [
 
   { cmd:  "is the following vehicle registered in qld",
     log:  (query) => "crawling qld government for the rego " + query,
-    func: (query) => {
-      return (query != "")
-        ? rego.search(query)
+    func: (query) =>
+       (query != "")
+        ? crawls.findrego(query)
         : Promise.resolve("Please enter a registration number")
-    }
   }
 ]
